+++ lib/src/fictional_progress.dart
@@ -1,24 +1,29 @@
 import 'dart:async';
 
-import 'debug_helpers.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
+import 'package:parent_progress/src/child_progress.dart';
 
-/// A class that simulates a progress. This is not a real progress tracker,
+import 'debug_helpers.dart';
+
+/// A class that simulates progress. This is not a real progress tracker,
 /// but rather an estimation tool that simulates progress based on predefined sizes and rates.
 ///
 /// This class manages progress calculation and notifies listeners about the
 /// current progress using a [ValueNotifier]. It uses a list of sizes to simulate
 /// the progress of data processing over time without actually processing any data.
-class FictionalProgress {
+class FictionalProgress extends ChildProgress {
   int _percentage = 0;
   final List<int>
       _sizes; // List of sizes that represent the segments of data or tasks to be simulated.
   int _totalSize = 0;
   Timer? _timer;
   double _processedSize = 0.0;
-  ValueNotifier<int> percentageNotifier;
+  ValueNotifier<double> processedSizeNotifier;
   Completer<void>? _completer;
-  String? uniqueName;
+  late double processingRatePerS;
+  late int _updateIntervalMs;
+  late double _targetSize;
+  bool isShowDebugPeriodicUpdate;
 
   /// Creates a [FictionalProgress] with a list of sizes.
   ///
@@ -29,8 +34,11 @@ class FictionalProgress {
   ///
   /// The [uniqueName] parameter can be used to uniquely identify the instance
   /// for debugging purposes.
-  FictionalProgress(this._sizes, {this.uniqueName})
-      : percentageNotifier = ValueNotifier<int>(0) {
+  FictionalProgress(
+    this._sizes, {
+    super.uniqueName,
+    this.isShowDebugPeriodicUpdate = false,
+  }) : processedSizeNotifier = ValueNotifier<double>(0) {
     _totalSize = _sizes.reduce((a, b) => a + b);
   }
 
@@ -42,6 +50,17 @@ class FictionalProgress {
 
   /// Returns the total size.
   int get getTotalSize => _totalSize;
+
+  /// Returns a list of sizes, where each size represents a stage of processing.
+  List get getSizes => _sizes;
+
+  // Setter method: sets the value of _updateIntervalMs and restarts the timer if it is already running.
+  set setUpdateIntervalMs(int value) {
+    _updateIntervalMs = value;
+    if (_timer != null) {
+      _restartTimer();
+    }
+  }
 
   /// Returns the current process index level based on the processed size.
   ///
@@ -68,37 +87,56 @@ class FictionalProgress {
     required double processingRatePerS,
     required int updateIntervalMs,
   }) async {
-    _timer?.cancel(); // Cancel the previous timer
-    _timer = null;
+    _stopTimer(); // Cancel the previous timer
     _completer?.complete(); // Complete the previous operation
     _completer = Completer<void>();
-
-    double targetSize = 0;
+    this.processingRatePerS = processingRatePerS;
+    _updateIntervalMs = updateIntervalMs;
+    _targetSize = 0;
     for (int i = 0; i <= processIndexLevel; i++) {
-      targetSize += _sizes[i];
+      _targetSize += _sizes[i];
     }
     printDebugInfo(
         "FictionalProgress ${uniqueName != null ? '${uniqueName!}: ' : ''}Start method started with : $_percentage");
-    _timer = Timer.periodic(Duration(milliseconds: updateIntervalMs), (timer) {
-      if (_processedSize < targetSize) {
-        _processedSize += processingRatePerS * (updateIntervalMs / 1000);
-        if (_processedSize > targetSize) {
-          _processedSize = targetSize;
+    _startTimer();
+    await _completer?.future;
+    _completer = null;
+  }
+
+  /// Starts the timer to simulate the progress.
+  void _startTimer() {
+    _timer = Timer.periodic(Duration(milliseconds: _updateIntervalMs), (timer) {
+      if (_processedSize < _targetSize) {
+        _processedSize += processingRatePerS * (_updateIntervalMs / 1000);
+        if (_processedSize > _targetSize) {
+          _processedSize = _targetSize;
         }
         _percentage = ((_processedSize / _totalSize) * 100).round();
         percentageNotifier.value = _percentage;
-        printDebugInfo(
-            "FictionalProgress ${uniqueName != null ? '${uniqueName!}: ' : ''}Processing size: $_processedSize, Percentage: $_percentage");
+        processedSizeNotifier.value = _processedSize;
+        if (isShowDebugPeriodicUpdate) {
+          printDebugInfo(
+              "FictionalProgress ${uniqueName != null ? '${uniqueName!}: ' : ''}Processing size: $_processedSize, Percentage: $_percentage");
+        }
       } else {
         printDebugInfo(
             "FictionalProgress ${uniqueName != null ? '${uniqueName!}: ' : ''}Level Finished and percentage reached: $_percentage");
-        _timer?.cancel();
-        _timer = null;
+        _stopTimer();
         _completer?.complete();
       }
     });
-    await _completer?.future;
-    _completer = null;
+  }
+
+  /// Restarts the timer.
+  void _restartTimer() {
+    _timer?.cancel();
+    _startTimer();
+  }
+
+  /// Stops the timer.
+  void _stopTimer() {
+    _timer?.cancel();
+    _timer = null;
   }
 
   /// Resets the progress and optionally initializes new sizes.
@@ -106,8 +144,7 @@ class FictionalProgress {
   /// If [newSizes] is provided, it will replace the current sizes and
   /// reinitialize the total size.
   void resetProgress({List<int>? newSizes}) {
-    _timer?.cancel(); // Cancel the previous timer
-    _timer = null;
+    _stopTimer(); // Cancel the previous timer
     _completer?.complete();
     _percentage = 0;
     _processedSize = 0.0;
@@ -121,13 +158,15 @@ class FictionalProgress {
           "FictionalProgress ${uniqueName != null ? '${uniqueName!}: ' : ''}New Sizes List Initialized: ${newSizes.join(', ')}");
     }
     percentageNotifier.value = _percentage;
+    processedSizeNotifier.value = _processedSize;
   }
 
   /// Disposes the progress and releases all resources.
   void dispose() {
-    _timer?.cancel();
+    _stopTimer();
     _completer?.complete();
     percentageNotifier.dispose();
+    processedSizeNotifier.dispose();
     printDebugInfo(
         "FictionalProgress ${uniqueName != null ? '${uniqueName!}: ' : ''}FictionalProgress disposed.");
   }